Center Issue Fixed (Thanks Eric)
I fixed my centering issue and added a few things. Actually, everything I added were buttons on the "Gym" page. The goal is to have the gym buttons expand when clicked on which can give you a more detailed layout of the gym; including highlights from the week; logged climbs; pictures or videos of routes (if available); and highlights from followers as well (if desired). From here, you can obviously add stats as well, log climbs, training notes, diet notes, or other (say if you just had a busy week or something of the sort). The grades climbed will be rated exponentially (2^y). I feel this is fair and maybe logical approach, but also most likely flawed but can be revisited at another time.

On the highlights page it will give you weekly highlights of the climbs you did as well as the changes (if any) that were logged to keep you in the loop of what you are doing right or wrong (ie is training improving your goals or not; is diet poor or good this week; are you projecting more this week; etc...)

The tricky part is adding more gyms... I have to figure out a way to add divs to the code so more gyms will be added. I plan on having a side scrolling feature eventually for the gyms that dont fit on the page.